computershop/requirements.txt
asgiref==3.6.0
Django==4.1.6
django_debug_toolbar==3.8.1
Pillow==7.2.0
pytz==2022.1
sqlparse==0.4.3